move v0 handlers into one single file
