--- jormungandr/src/rest/v0/block/next_id.rs
@@ -1,43 +1,0 @@
-use super::parse_block_hash;
-
-use chain_storage::store;
-
-use actix_web::error::{Error as ActixError, ErrorBadRequest, ErrorInternalServerError};
-use actix_web::{Path, Query, State};
-use blockchain::BlockchainR;
-use bytes::Bytes;
-
-pub fn handle_request(
-    blockchain: State<BlockchainR>,
-    block_id_hex: Path<String>,
-    query_params: Query<QueryParams>,
-) -> Result<Bytes, ActixError> {
-    let block_id = parse_block_hash(&block_id_hex)?;
-    // FIXME
-    // POSSIBLE RACE CONDITION OR DEADLOCK!
-    // Assuming that during update whole blockchain is write-locked
-    // FIXME: don't hog the blockchain lock.
-    let blockchain = blockchain.lock_read();
-    let storage = blockchain.storage.read().unwrap();
-    store::iterate_range(&*storage, &block_id, &blockchain.get_tip().unwrap())
-        .map_err(|e| ErrorBadRequest(e))?
-        .take(query_params.get_count())
-        .try_fold(Bytes::new(), |mut bytes, res| {
-            let block_info = res.map_err(|e| ErrorInternalServerError(e))?;
-            bytes.extend_from_slice(block_info.block_hash.as_ref());
-            Ok(bytes)
-        })
-}
-
-const MAX_COUNT: usize = 100;
-
-#[derive(Deserialize)]
-pub struct QueryParams {
-    count: Option<usize>,
-}
-
-impl QueryParams {
-    pub fn get_count(&self) -> usize {
-        self.count.unwrap_or(1).min(MAX_COUNT)
-    }
-}
