+++ jormungandr/src/rest/v0/message/logs.rs
@@ -1,7 +1,7 @@
 use crate::fragment::Logs;
-use actix_web::{App, Json, Responder, State};
-use futures::Future;
+use actix_web::App;
 use std::sync::{Arc, Mutex};
+use crate::rest::v0::handlers;
 
 pub fn create_handler(
     logs: Arc<Mutex<Logs>>,
@@ -10,12 +10,6 @@ pub fn create_handler(
         let app_prefix = format!("{}/v0/fragment/logs", prefix);
         App::with_state(logs.clone())
             .prefix(app_prefix)
-            .resource("", |r| r.get().with(handle_request))
+            .resource("", |r| r.get().with(handlers::get_message_logs))
     }
-}
-
-fn handle_request(logs: State<Arc<Mutex<Logs>>>) -> impl Responder {
-    let logs = logs.lock().unwrap();
-    let logs = logs.logs().wait().unwrap();
-    Json(logs)
 }
