+++ jormungandr/src/rest/v0/mod.rs
@@ -4,3 +4,5 @@ pub mod message;
 pub mod node;
 pub mod tip;
 pub mod utxo;
+
+pub mod handlers;
